# perf(ext/streams): fast path when consuming body of tee'd stream