Add a fast path for consuming the body of cloned `Request`/`Response`, which is very common specially when using `cache` API.

**patch - deno_post_bin**

```
Running 20s test @ http://127.0.0.1:4558/
  138683 requests in 20.10s, 19.05MB read
Requests/sec:   6899.70
Transfer/sec:      0.95MB
```
**patch - deno_post_json**
```
Running 20s test @ http://127.0.0.1:4558/
  312885 requests in 20.10s, 42.37MB read
Requests/sec:  15566.42
Transfer/sec:      2.11MB
```

**main - deno_post_bin**
```
Running 20s test @ http://127.0.0.1:4558/
  2 threads and 10 connections
  107811 requests in 20.00s, 14.81MB read
Requests/sec:   5389.68
Transfer/sec:    757.92KB
```

**main - deno_post_json**
```
Running 20s test @ http://127.0.0.1:4558/
  2 threads and 10 connections
  246295 requests in 20.00s, 33.35MB read
Requests/sec:  12313.82
Transfer/sec:      1.67MB
```


Modified the following for benching

https://github.com/denoland/deno/blob/9df8d9d83140beb24a53b0ac52b1eaf2ea34c095/cli/bench/http/deno_post_json.js#L13

```js
        // const json = await request.json();
        const clone = request.clone();
        const [json] = await Promise.all([
          request.json(),
          clone.json(),
        ]);
```