@marcosc90 2022-10-23
> I don't think this works. If the A side reads the stream, the B side gets nothing because the underlying resource has been read now.

To avoid that, I placed the fast path inside `ReadableByteStreamController.pull`, so `controller.enqueue` can be used so the read chunks are enqueued correctly to the other branch.

https://github.com/denoland/deno/blob/d65f4b3d99dd3b0aa4d1e51027c9381c15187238/ext/web/06_streams.js#L684-L696

All the tests I've done work as expected so does all WPT.